Ensure mapsforge maps are valid before creating an offline map entry
