--- app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/infrastructure/vector_maps/attribution/OfflineMapAttributionExtractor.kt
@@ -1,5 +1,0 @@
-package com.kylecorry.trail_sense.tools.offline_maps.infrastructure.vector_maps.attribution
-
-interface OfflineMapAttributionExtractor {
-    suspend fun getAttribution(path: String): String?
-}
