--- app/src/main/java/com/kylecorry/trail_sense/tools/offline_maps/infrastructure/vector_maps/bounds/OfflineMapBoundsCalculator.kt
@@ -1,7 +1,0 @@
-package com.kylecorry.trail_sense.tools.offline_maps.infrastructure.vector_maps.bounds
-
-import com.kylecorry.sol.science.geology.CoordinateBounds
-
-interface OfflineMapBoundsCalculator {
-    suspend fun getBounds(path: String): CoordinateBounds?
-}
